Loading input data
This code uploads the flight schedule at Heathrow, for 2017, onto Python.
The flight schedule is processed to extract the data for Terminal 5 (the focus of the project) and to remove some unnecessary information.

The code then extends to calculate the turnaround times (the time an aircraft needs to remain at the airport between its arrival and departure) for each flight. The flight schedule is re-formatted to include this information and returned as a "Turnarounds Table" CSV file.