Performs the "Stand Allocation Heuristic"
There were 2 main functions that there defined in this class. The "Daily_Sched" and "Monthly_Sched" produced the daily and monthly allocation schedules for the flights, respectively.
The ‘Daily_Sched’ function took the day of the date and a string input. It then ran the ‘Schedule_F’, ‘Schedule_E’, ‘Schedule_D’ and ‘Schedule_C’ functions, from the other classes, in this order. The largest aircraft were assigned first and smallest last. The completed flight turnarounds table, with the stand allocations, was then saved as a .CSV file.
The ‘Monthly_Sched’ function took the numerical value of the month as an input and used it to extract the flight schedule data for that month. Then, the function looped over the dataset, to create a turnarounds table and apply the ‘Daily_Sched’ function to every day of the month. The allocation schedules for each day of the month were then saved as CSV files in a folder, labelled by the month.